refactor: eliminate redundant parameter type mapping
Fixed the root cause of parameter type override issues by removing
redundant type mapping in MapQueryColumns. Parameters are already
properly mapped by the query analyzer using database introspection,
so re-mapping them caused the int override for limit/offset to be lost.

Changes:
- Renamed MapQueryColumns to MapResultColumns to clarify it only maps
  SELECT result columns, not parameters
- Removed redundant parameter type mapping (lines 260-274)
- Fixed codegen to use range by index to avoid value copies
- Updated service layer to use int instead of int32 for limit parameters
- All integration tests now pass with proper int type for limit/offset

The generated code now correctly uses int for limit/offset parameters
throughout the entire stack: generated queries, repository layer, and
service layer.

Resolves #43

+++ example-app/service/post_service.go
@@ -26,7 +26,7 @@ func NewPostService(postRepo PostRepository) *postService {
 // The service layer focuses on business logic, not data conversion
 
 func (s *postService) GetPublishedPosts(ctx context.Context, limit int) ([]domain.PostSummary, error) {
-	posts, err := s.postRepo.GetPublishedPosts(ctx, int32(limit))
+	posts, err := s.postRepo.GetPublishedPosts(ctx, limit)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get published posts: %w", err)
 	}
@@ -57,7 +57,7 @@ func (s *postService) GetUserPosts(ctx context.Context, userID uuid.UUID) ([]dom
 }
 
 func (s *postService) GetPostsWithStats(ctx context.Context, limit int) ([]domain.PostWithStats, error) {
-	posts, err := s.postRepo.GetPostsWithStats(ctx, int32(limit))
+	posts, err := s.postRepo.GetPostsWithStats(ctx, limit)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get posts with stats: %w", err)
 	}
